Fix debug mode set and get
Fix debug const. Now managed via utils, as part of addon preferences and the constant, and will return whichever one is true, otherwise false. Relying on importing from const still only provided us a copy of the variable and changing it with global only changed the copy of the DEBUG variable on that specific file, and it did not propagate that change globally to the project.