Initial depsgraph optimisation pass
Defer the cleanup to slightly after the update. This isn't enough, and so we check first if scene object count have changed, only then can we proceed with the cleanup.

We do the scene object check outside the actual cleanup function as we may want to trigger it regardless from other places and so we can keep it flexible.

While it still runs in the main thread, the optimisations made to the new method allow it to be much more efficient and hopefully less noticeable when it runs.

Additionally, it is possible to expand this to only run after x seconds have passed after the last cleanup, though that is less desirable since we need to keep the references updated and cleaned up.