Direct data assign for legacy object sets
Test a direct data assign to test data assignment performance. Reduce checks and updates and focus on simply copying the data over with the special assignment methods that have safeguards but are slow to iterate.

On another note, perhaps it would be good to revisit those methods to try and improve their performance as well.