Add more edge data conversion options
Add option to convert using edge weight maximum value at vertex instead of averaging the weight values.

Add options to choose to which channels the bevel/crease conversions are applied to.

Converting to vertex color is now restricted to either convert bevel weights or creases to vertex colours. Not both at the same time, for an easier time dealing with clashing priorities.