Implement Object Sets auto-reassign on duplication
Leverage the object-level properties feature that carries over properties for duplicated objects.
When objects are duplicated, use the object-level properties to track and read what were the original object sets the object(s) belong to and automatically assign them.
This allows the user to keep working and alleviating some more manual and tedious Object Set assignment for duplicated objects.